Busy-wait implementation works now
probably because it actually does the busy wait now, as opposed to
checking once for non-empty/non-full. AFAICT, this was happening:

- buffer is empty().
- A, B enter remove, and must both wait on itemAvailable_.
- C enters insert; buffer is not full(), so it inserts and signal()s on
  itemAvailable_.
- A, B are now unblocked.
- A acquire()s, removes item, and release()s.
- B acquire()s, removes item...
- ...oops.

with the busy wait, B must check again that the buffer is not empty();
this check fails, and so B waits again on itemAvailable_.

this understanding assumes that uCondLock::signal() frees *all*
waiting tasks, perhaps by promoting them from a "blocked" queue to a
"ready" queue...but it doesn't, so i'm a bit confused. maybe this is
because itemAvailable_, slotFree_ share their underlying uOwnerLock? but
it makes sense for them to, because we've got to synchronize updates to
size_ and head_. if we used head_, tail_ instead, updated head_ in
remove() and tail_ in insert(), padded the circular buffer with an
additional "sentry" slot to distinguish empty(), full() cases, and then
had two separate uOwnerLocks, this might work...but there would need to
be stricter checking for full(), empty() conditions, and i'm not
convinced it's worth it.

another observation: uCondLock isn't *necessarily* tied to a condition;
you can have a task wait on it whenever you want. if, for instance, we
dropped the empty()/full() checks altogether, we'd have chaos - which
seems obvious now that i type it out :)

so yeah: i mostly understand this, but there's that weirdness with
signal() vs. broadcast() in uCondLock.