Implementation of Sequencer from A3Q2
- added semaphore-synchronized sequencer; we use a semaphore with bound
  1 here to synchronize access to all methods. (i'm not convinced that
  read access really needs to be synchronized in the same way; maybe we
  can get rid of that?)
- using the Sequencer from q2Main as a quick test. i originally wrote
  the Consumer task without yield(), but found upon running it that each
  task would loop through all 10 ticket()s before the next task could be
  created (either that, or the task must yield() control before
  something else is run? i'll have to see if there's preemptive control
  passing in uC++...) anyways, with yield() the tasks correctly grab
  consecutive and monotonic integers even while interleaving control.
- started an implementation of EventCounter. no synchronization yet.
- ran into a weird "symbol already defined" error when i tried to define
  the functions in Sequencer outside the class, as i had done for the
  A3Q1 classes. i'll have to look closer to see why this is happening.

I was originally confused about the role of advance(), since I had
misinterpreted the EventCounter to *be* a server as opposed to being a
tool *used by* the server. With that understanding, the motivation for
advance() becomes clear: "servers" (in this case, the
EventCounter/Sequencer-based BoundedBuffer) force clients to wait using
the E.await(S.ticket()) idiom, perform their work, then admit another
client using E.advance().

The FIFO behavior of EventCounter, then, is essentially providing a
condition lock/variable; the shared counter underneath provides some
manner of synchronization (I think) so that the new BoundedBuffer
implementation can very closely mirror the existing
uOwnerLock/uCondLock-based implementations (which is probably the point
of this exercise.)